LayoutTests/imported/w3c/web-platform-tests/preload/onload-event-expected.txt

CONSOLE MESSAGE: <link rel=preload> must have a valid `as` value
CONSOLE MESSAGE: <link rel=preload> cannot have the empty string as `as` value

PASS Makes sure that preloaded resources trigger the onload event
